fix error in logging

+++ libraries/Logger/Logger.h
@@ -36,9 +36,9 @@ class Logger {
     File m_file;
     String m_filename;
     unsigned long int m_nb_entry;
-    unsigned int m_last_log=0;
-    unsigned int m_last_xbee=0;
-    unsigned int m_time = 0;
+    unsigned long m_last_log=0;
+    unsigned long m_last_xbee=0;
+    unsigned long m_time = 0;
 
     Sailboat* m_boat = NULL;
 };
